adapt to sk

+++ src/dpss/matrix_preparer.py
@@ -1,6 +1,7 @@
 import gzip
 import logging
 import os
+from pathlib import Path
 import warnings
 
 from more_itertools import one
@@ -30,13 +31,13 @@ class MatrixPreparer:
 
     hca_zipped_filenames = {
         'genes': 'genes.tsv.gz',
-        'barcodes': 'cells.tsv.gz',
+        'barcodes': 'barcodes.tsv.gz',
         'matrix': 'matrix.mtx.gz'
     }
 
     hca_filenames = {
         'genes': 'genes.tsv',
-        'barcodes': 'cells.tsv',
+        'barcodes': 'barcodes.tsv',
         'matrix': 'matrix.mtx'
     }
 
@@ -53,48 +54,94 @@ class MatrixPreparer:
         'genes': ['featurekey', 'featurename']
     }
 
+    @staticmethod
+    def translate_filename(fname):
+        if fname in ['cells.tsv', 'barcodes.tsv']:
+            return 'barcodes.tsv'
+        elif fname in ['genes.tsv', 'features.tsv']:
+            return 'genes.tsv'
+        elif fname == 'matrix.mtx':
+            return 'matrix.mtx'
+        raise ValueError(f'Couldn\'t understand filename {fname}')
+
     def __init__(self, mtx_info: MatrixInfo):
         self.info = mtx_info
 
-    def unzip(self) -> None:
+    def unzip(self) -> List[MatrixInfo]:
         """
         Extract files from top-level zip archive, uncompress .gz files, and
         remove archive.
         """
         log.info(f'Unzipping {self.info.zip_path}')
-        os.mkdir(self.info.extract_path)
+
         with ZipFile(self.info.zip_path) as zipfile:
-            for member in zipfile.namelist():
-                filename = os.path.basename(member)
-                source = zipfile.open(member)
-                target = open(os.path.join(self.info.extract_path, filename), 'wb')
-                with source, target:
-                    shutil.copyfileobj(source, target)
+            zipfile.extractall(self.info.extract_path)
         os.remove(self.info.zip_path)
 
-        extracted_files = os.listdir(self.info.extract_path)
-        assert all(filename in extracted_files for filename in self.hca_zipped_filenames.values())
+        new_infos = []
 
         with DirectoryChange(self.info.extract_path):
-            for gzfilename in self.hca_zipped_filenames.values():
-                with gzip.open(gzfilename, 'rb') as gzfile:
-                    filename = remove_ext(gzfilename, '.gz')
-                    with open(filename, 'wb') as outfile:
-                        shutil.copyfileobj(gzfile, outfile)
+            for mtx_dir in os.listdir('.'):
+                if not os.path.isdir(mtx_dir):
+                    log.info(f'Skipping non-matrix entry {mtx_dir}')
+                    continue
+                with DirectoryChange(mtx_dir):
+                    for gzfilename in os.listdir('.'):
+                        with gzip.open(gzfilename, 'rb') as gzfile:
+                            filename = remove_ext(gzfilename, '.gz')
+                            with open(filename, 'wb') as outfile:
+                                shutil.copyfileobj(gzfile, outfile)
                         os.remove(gzfilename)
+            new_infos.append(MatrixInfo(
+                zip_path=None,
+                extract_path=os.path.join(self.info.extract_path, mtx_dir),
+                source=self.info.source,
+                project_uuid=self.info.project_uuid,
+                lib_con_approaches=self.info.lib_con_approaches
+            ))
+
+        return new_infos
 
     def preprocess(self):
         """
         Extract gzip files and transform for ScanPy compatibility.
         """
         with DirectoryChange(self.info.extract_path):
-            for filekey, filename in self.hca_filenames.items():
-                log.info(f'Processing file {filename}')
-                if filename.endswith('.tsv'):
-                    self._preprocess_tsv(filename, self.scanpy_tsv_columns[filekey])
-                elif filename.endswith('.mtx'):
-                    self._preprocess_mtx(filename)
-                os.rename(filename, self.scanpy_filenames[filekey])
+
+            header, mtx = self._read_matrixmarket(self.hca_filenames['matrix'])
+            mtx.iloc[:, 2] = np.rint(mtx.iloc[:, 2]).astype(np.int32)
+            self._write_matrixmarket(self.hca_filenames['matrix'], header, mtx)
+
+            rewrite = False
+
+            genes = self._read_tsv(self.hca_filenames['genes'], header=False)
+            genes = genes.applymap(str)
+
+            if genes.shape[0] == mtx.iloc[0, 0] + 1:
+                # header
+                rewrite = True
+                genes = genes.iloc[1:, :]
+
+            if genes.shape[1] < 2:
+                genes[1] = genes.iloc[:, 0]
+                rewrite = True
+
+            if rewrite:
+                self._write_tsv(self.hca_filenames['genes'], genes)
+
+            rewrite = False
+
+            barcodes = self._read_tsv(self.hca_filenames['barcodes'], header=False)
+            if barcodes.shape[0] == mtx.iloc[0, 1] + 1:
+                # header
+                rewrite = True
+                barcodes = barcodes.iloc[1:, :]
+
+            if rewrite:
+                self._write_tsv(self.hca_filenames['barcodes'], barcodes)
+
+            for filename in self.hca_filenames.values():
+                os.rename(filename, self.translate_filename(filename))
 
     def prune(self, keep_frac: float) -> None:
         """
@@ -236,8 +283,11 @@ class MatrixPreparer:
         :param keep_cols: columns to keep.
         """
         df = cls._read_tsv(filename, True)
-        assert all(col in df.columns for col in keep_cols)
-        df = df[keep_cols]
+        if all(col in df.columns for col in keep_cols):
+            df = df[keep_cols]
+        else:
+            log.warning('Ignoring named columns')
+            df = df.iloc[:, :2]
         cls._write_tsv(filename, df)
 
     @classmethod
